Submission状態遷移をCASで実装

+++ src/ComputeSharp/Graphics/Pipelines/SubmissionState.cs
@@ -1,6 +1,6 @@
 namespace ComputeSharp.Graphics.Pipelines;
 
-internal enum SubmissionState : byte
+internal enum SubmissionState : int
 {
     Reserved = 0,
     Recording = 1,
